requirements.txt
openai==0.27.8
pyfzf==0.3.1
python-dotenv==1.0.0
rich==13.4.2
